Clear data on server wipe
This change properly fixes an older issue where the data file would desync and cause players to be unable to spawn their mini whenever the server wipes.

This also improves the behavior when a desync does occur by simply spawning the mini instead of just showing an error and requiring that the player run the command again

Cooldowns now also reset on server wipe as expected.

Misc changes:
- Cleaned up some if statements
- Replaced usage of the "SpawnMini" string with the plugin Name field