Improve cooldown logic
The time saved in the data file is now the last time that the mini was spawned, not the projected end of the cooldown. This allows for changes to cooldown permission to immediately reduce existing cooldowns which will solve multiple issues people have reported.

Also improved the display of the cooldown duration to be more human friendly.